requirements.txt
trimesh==3.6.20
numpy==1.18.1
matplotlib==3.1.3
scikit_image==0.16.2
ipdb==0.12.3
yacs==0.1.8
open3d==0.10.0.0
Cython==0.29.21
scikit_learn==0.24.2
tensorboardX==2.2
opencv_python==4.4.0.46
albumentations==0.5.1 --no-binary=imgaug,albumentations
ray>=1.0.1
einops==0.3.0
kornia==0.4.1
loguru==0.5.3
tqdm
autopep8
pylint
ipython
jupyterlab
h5py==3.1.0
pytorch-lightning==1.3.5
joblib>=1.0.1
pykdtree
torchmetrics==0.5.1
terminaltables==3.1.0
cvxpy==1.2.0
